Parallel refactor main to use the new HTTPServer
