+++ src/main/java/io/github/christophermanahan/carnitas/Connection.java
@@ -1,9 +1,7 @@
 package io.github.christophermanahan.carnitas;
 
 public interface Connection extends AutoCloseable, Readable {
-    Receiver receiver();
-
-    void send(Response response);
+    void send(HTTPResponse response);
 
     void close();
 }
