+++ src/main/java/io/github/christophermanahan/carnitas/Handler.java
@@ -1,5 +1,5 @@
 package io.github.christophermanahan.carnitas;
 
 public interface Handler {
-    Response handle(Request request);
+    HTTPResponse handle(HTTPRequest request);
 }
