--- src/main/java/io/github/christophermanahan/carnitas/Handler2.java
@@ -1,5 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface Handler2 {
-    HTTPResponse2 handle(HTTPRequest2 request);
-}
