+++ src/main/java/io/github/christophermanahan/carnitas/Parser.java
@@ -3,5 +3,5 @@ package io.github.christophermanahan.carnitas;
 import java.util.Optional;
 
 public interface Parser {
-    public Optional<Request> parse(Receiver receiver);
+    Optional<HTTPRequest> parse(Reader reader);
 }
