--- src/main/java/io/github/christophermanahan/carnitas/Parser2.java
@@ -1,7 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-import java.util.Optional;
-
-public interface Parser2 {
-    Optional<HTTPRequest2> parse(Reader reader);
-}
