--- src/main/java/io/github/christophermanahan/carnitas/Receiver.java
@@ -1,6 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface Receiver {
-    public String receiveLine();
-    public String receiveCharacters(int amount);
-}
