--- src/main/java/io/github/christophermanahan/carnitas/Request.java
@@ -1,6 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface Request {
-    public RequestMethod requestMethod();
-    public String body();
-}
