--- src/main/java/io/github/christophermanahan/carnitas/RequestBuilder.java
@@ -1,7 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface RequestBuilder {
-    public RequestBuilder requestMethod(String requestMethod);
-    public RequestBuilder body(String body);
-    public Request build();
-}
