--- src/main/java/io/github/christophermanahan/carnitas/RequestMethod.java
@@ -1,7 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public enum RequestMethod {
-    GET,
-    HEAD,
-    POST
-}
