+++ src/main/java/io/github/christophermanahan/carnitas/RequestParser.java
@@ -4,83 +4,67 @@ import java.util.List;
 import java.util.Optional;
 
 public class RequestParser implements Parser {
-    private final RequestBuilder requestBuilder;
-    private Receiver receiver;
-    private boolean receiving = true;
-    private String requestMethod;
-    private int contentLength = -1;
-    private String body = "";
-
-    RequestParser(RequestBuilder requestBuilder) {
-       this.requestBuilder = requestBuilder;
+    public Optional<HTTPRequest> parse(Reader reader) {
+        Optional<String> method = getMethod(reader);
+        Optional<Integer> contentLength = parseHeadersForContentLength(reader);
+        Optional<String> body = getBody(reader, contentLength);
+        return request(method, body);
     }
 
-    public Optional<Request> parse(Receiver receiver) {
-        this.receiver = receiver;
-        resetState();
-        try {
-            parseStatusLine(receiver.receiveLine());
-            parseRequest();
-            return Optional.of(buildRequest());
-        } catch (RuntimeException e) {
-            return Optional.empty();
+    private Optional<String> readLine(Reader reader) {
+        return reader.readUntil(Constants.CRLF);
+    }
+
+    private Optional<String> getMethod(Reader reader) {
+        return readLine(reader)
+          .map(this::extractMethod);
+    }
+
+    private String extractMethod(String statusLine) {
+        return List.of(statusLine.split(" ")).get(0);
+    }
+
+    private Optional<Integer> parseHeadersForContentLength(Reader reader) {
+        Optional<Integer> contentLength = Optional.empty();
+        Optional<String> header;
+        while((header = header(reader)).isPresent()) {
+            if (isContentLength(header)) {
+                contentLength = extractContentLength(header);
+            }
         }
+        return contentLength;
     }
 
-    private void resetState() {
-        this.receiving = true;
-        this.contentLength = -1;
-        this.body = "";
+    private Optional<String> header(Reader reader) {
+        return readLine(reader)
+          .filter(this::isNotEmpty);
     }
 
-    private void parseStatusLine(String statusLine) {
-        requestMethod = List.of(statusLine.split(" ")).get(0);
+    private boolean isNotEmpty(String header) {
+        return !header.isEmpty();
     }
 
-    private void parseRequest() {
-        while (receiving) {
-            parseLine(receiver.receiveLine());
-        }
+    private boolean isContentLength(Optional<String> header) {
+        return header
+          .filter(s -> s.contains(Headers.CONTENT_LENGTH))
+          .isPresent();
     }
 
-    private void parseLine(String received) {
-        getContentLength(received);
-        getBody(received);
-        endRequest(received);
+    private Optional<Integer> extractContentLength(Optional<String> contentLength) {
+        return contentLength
+          .map(s -> s.substring(Headers.CONTENT_LENGTH.length()))
+          .map(Integer::parseInt);
     }
 
-    private void getContentLength(String received) {
-        if (isPostWithBody(received)) {
-            contentLength = Integer.parseInt(received.substring(Headers.CONTENT_LENGTH.length()));
-        }
+    private Optional<String> getBody(Reader reader, Optional<Integer> contentLength) {
+        return contentLength
+          .filter(length -> length > 0)
+          .flatMap(reader::read);
     }
 
-    private boolean isPostWithBody(String received) {
-        return RequestMethod.valueOf(requestMethod) == RequestMethod.POST && received.contains(Headers.CONTENT_LENGTH);
-    }
-
-    private void getBody(String received) {
-        if (contentLength > 0 && received.isEmpty()) {
-            body = receiver.receiveCharacters(contentLength);
-//            body = receiver.receiveLine();
-            stopReceiving();
-        }
-    }
-
-    private void endRequest(String received) {
-        if (received.isEmpty()) {
-            stopReceiving();
-        }
-    }
-
-    private void stopReceiving() {
-        receiving = false;
-    }
-
-    private Request buildRequest() {
-        return requestBuilder
-          .requestMethod(requestMethod)
-          .body(body)
-          .build();
+    private Optional<HTTPRequest> request(Optional<String> method, Optional<String> body) {
+        return method
+          .map(HTTPRequest::new)
+          .map(request -> request.withBody(body));
     }
 }
