--- src/main/java/io/github/christophermanahan/carnitas/Response.java
@@ -1,5 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface Response {
-    byte[] serialize();
-}
