--- src/main/java/io/github/christophermanahan/carnitas/ResponseBuilder.java
@@ -1,8 +1,0 @@
-package io.github.christophermanahan.carnitas;
-
-public interface ResponseBuilder {
-    public ResponseBuilder statusCode(String statusCode);
-    public ResponseBuilder version(String version);
-    public ResponseBuilder body(String body);
-    public Response build();
-}
