+++ src/test/java/io/github/christophermanahan/carnitas/HTTPRequestTest.java
@@ -1,20 +1,31 @@
 package io.github.christophermanahan.carnitas;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class HTTPRequestTest {
     @Test
-    void getsRequestMethod() {
-        RequestMethod requestMethod = RequestMethod.GET;
-        Assertions.assertEquals(requestMethod, new HTTPRequest(requestMethod, null).requestMethod());
+    void itHasAMethod() {
+        String method = "GET";
+        HTTPRequest request = new HTTPRequest(method);
+
+        String requestMethod = request.method();
+
+        assertEquals(method, requestMethod);
     }
 
     @Test
-    void getsBody() {
-        String body = "name=<something>";
-        Assertions.assertEquals(body, new HTTPRequest(null, body).body());
+    void itHasABody() {
+        String method = "GET";
+        Optional<String> body = Optional.of("name=<something>");
+        HTTPRequest request = new HTTPRequest(method)
+          .withBody(body);
+
+        Optional<String> requestBody = request.body();
+
+        assertEquals(body, requestBody);
     }
 }
